# Use a variable-size array to represent ordered labels in maps