--- a/sdk/metric/sdk.go
+++ b/sdk/metric/sdk.go
@@ -664,6 +664,10 @@ func (m *SDK) checkpoint(ctx context.Context, descriptor *export.Descriptor, rec
 	}
 	recorder.Checkpoint(ctx, descriptor)
 
+	// TODO Labels are encoded once per collection interval,
+	// instead of once per bound instrument lifetime.  This can be
+	// addressed similarly to OTEP 78, see
+	// https://github.com/jmacd/opentelemetry-go/blob/8bed2e14df7f9f4688fbab141924bb786dc9a3a1/api/context/internal/set.go#L89
 	exportLabels := export.NewLabels(labels.slice, m.labelEncoder.Encode(labels.slice), m.labelEncoder)
 	exportRecord := export.NewRecord(descriptor, exportLabels, recorder)
 	err := m.batcher.Process(ctx, exportRecord)