Rather than using a label encoder to encode a unique representation of the label set, we can use an `interface{}` containing an array value of the appropriate size.  (This was described as "use dangerous reflection black magic" in the Metrics SIG meeting notes today.) This improve benchmarks significantly. The code has been hand-written for up to 10 labels to avoid the overhead of `reflect`.

Before:

```
goos: darwin
goarch: amd64
pkg: go.opentelemetry.io/otel/sdk/metric
BenchmarkLabels_1-8    	 4147288	       266 ns/op	      80 B/op	       2 allocs/op
BenchmarkLabels_2-8    	 2574391	       447 ns/op	      96 B/op	       2 allocs/op
BenchmarkLabels_4-8    	 1570831	       806 ns/op	     144 B/op	       2 allocs/op
BenchmarkLabels_8-8    	  833871	      1432 ns/op	     224 B/op	       2 allocs/op
BenchmarkLabels_16-8   	  445599	      2703 ns/op	     400 B/op	       2 allocs/op
```

After:
```
goos: darwin
goarch: amd64
pkg: go.opentelemetry.io/otel/sdk/metric
BenchmarkLabels_1-8    	 9424884	       124 ns/op	      96 B/op	       2 allocs/op
BenchmarkLabels_2-8    	 7479254	       163 ns/op	     144 B/op	       2 allocs/op
BenchmarkLabels_4-8    	 5067748	       236 ns/op	     240 B/op	       2 allocs/op
BenchmarkLabels_8-8    	 3274238	       359 ns/op	     432 B/op	       2 allocs/op
BenchmarkLabels_16-8   	  763551	      1377 ns/op	    1584 B/op	       3 allocs/op
```