@jmacd 2020-03-06
Note that use of an array-valued key makes map lookups more expensive, but this is still an overall performance win.

Before:
```
BenchmarkInt64CounterAdd-8                             	10995777	       108 ns/op	       0 B/op	       0 allocs/op
BenchmarkInt64CounterHandleAdd-8                       	72600636	        16.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkFloat64CounterAdd-8                           	10173048	       112 ns/op	       0 B/op	       0 allocs/op
BenchmarkFloat64CounterHandleAdd-8                     	63691292	        18.8 ns/op	       0 B/op	       0 allocs/op
```

After:
```
BenchmarkInt64CounterAdd-8                             	 7974315	       151 ns/op	       0 B/op	       0 allocs/op
BenchmarkInt64CounterHandleAdd-8                       	71714146	        16.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkFloat64CounterAdd-8                           	 7449758	       157 ns/op	       0 B/op	       0 allocs/op
BenchmarkFloat64CounterHandleAdd-8                     	63894693	        18.6 ns/op	       0 B/op	       0 allocs/op

```