@jmacd 2020-03-06
Note that this change removes the caching of encoded labels.  It's possible to regain this by adding support for the label set to cache the encoding for >=1 label encoder, instead. I did it that way in OTEP 78, see here:

https://github.com/jmacd/opentelemetry-go/blob/8bed2e14df7f9f4688fbab141924bb786dc9a3a1/api/context/internal/set.go#L89