@jmacd 2020-03-06
Another note: if the benchmark was to measure the cost of update-to-export of a single metric event, and we compare with previous "single encoding" optimization, this approach costs more overall because we're computing two encodings (one for the map and one for the export). I'm still happy with this direction because it reduces confusion about handling label encoders.