@jmacd 2020-03-06
We can slightly optimize this code as noted here: https://github.com/open-telemetry/opentelemetry-go/pull/523/files#diff-4c3de179542bac3c5ceacd2090215160R371

As there is a forthcoming OTEP proposing to remove LabelSet, I decided to leave several potential optimizations out.